+++ sketch/src/androidTest/java/com/github/panpf/sketch/test/utils/TestDisplayCountDisplayTarget.kt
@@ -1,0 +1,8 @@
+package com.github.panpf.sketch.test.utils
+
+import com.github.panpf.sketch.target.DisplayTarget
+
+class TestDisplayCountDisplayTarget : DisplayTarget {
+
+    override val supportDisplayCount: Boolean = true
+}